MSI: 440: Adapt DataProvider; 439: Adapt TermDropdownStrategy
